feat(server): remove cookieSession middleware
We're only saving the JWT we generate after login and send it as a cookie in the client.
cookieSession middleware is for sending 'session cookie' related specifically for session authentication and we don't need that in our case.

We just need to let the browser send the jwt cookie on every HTTP request and verify it using cookieParser