fix(auth.service): disable caching for fetching user
The reason we disable this is because Webkit automatically caches the previous http requests if we press back button when choosing between emails.

Try login with google, then press back without choosing an email to login with. Then see in the network panel that the request to /api/auth/user is cached