fix(server): blacklist logout route for csurf
Before, all routes are protected by csurf middleware, so we can't logout user if both jwt and refresh token are expired because it will return a 403 response error.

We need to blacklist /api/logout/route because when both jwt and refresh token expires we're going to force logout the user, so we disable the protection on that route.