refactor(client): update bookmarks$ observable stream
The previous one depends on `selectBookmarksLoading` selector w/c only
happens once on load.

this would mean that if a selector that depends on a slice of state outside the `bookmarks` state changes,
then the `withLatestFrom` operator won't run since `selectBookmarkLoading` won't run because the bookmarks
state is not changed.