refactor(client): update bookmarks$ observable
The problem for this change is because it flashes the "empty items
message" because of extra emission caused by clear* actions.

We replaced the combineLatest with withLatestFrom approach so we
only subscribed to the value of selectCurrentBookmarks once
selectBookmarksLoading emits a value w/c is NOT affected when clear*
action is dispatched

There are 2 container components that need different
approach to retrieve the bookmarks:
1. BookmarkSearchResultsContainerComponent
2. BookmarksByTagContainerComponent

These 2 components are visited multiple times and are reused
since Angular uses the same instance for same route navigation.

They change routes based on query params and params respectively.

The reason we need to have a different approach is we always clear
the bookmarks by dispatching clear* actions w/c causes another state
change causing the selector(selectCurrentBookmarks) to run again (try adding tap to check the
values).

combineLatest isn't the right operator in this case because, the extra
state change by clear* actions is intercepted and selectBookmarks
dispatches 1 unnecessary value emission. By using withLatestFrom
we are only subscribing to the value emitted by selectCurrentBookmarks only if
the selectLoading dispatches.