fix(server): add config file and setup dotenv
I have to correct the setup for configuration. Previously, i added
the config file in .gitignore.

For some reason, i decided to use config files instead of .env files.
In hindsight, this was a wrong decision.

To fix this what I did was:
* remove the config file in .gitignore
* setup dotenv package to inject values from .env files into the
environment (process.env)
* load only dotenv files when environment is NOT in production because
we'll manually insert environment values in PaaS (i.e heroku)
* setup npm scripts to run in different environments
* add a placeholder for .env files
  placeholder must be in suffixed with the desired environment

  For example:
  if you want to add a staging environment name the file as
  .env.staging