fix(server): add tentative fix for deployment error #3
Typescript compilation shows "cannot find module" errors
during compilation. For some reason, it still builds the files
to javascript BUT it breaks the deployment process to either
heroku or docker.

As a tentative solution, I decided to just suppress the typescript
errors related to the error since it still compiles it to javascript.