fix(client): conditionally dispatch saveBookmarkSuccess effect #10
We only need to display the newly saved bookmark if we're on /bookmarks
route.

Let's say we're saving a bookmark and it's taking it too long and
we do a search then while searching, the saving has finished. What
happens is the newly saved bookmark will be added to the search
results.

The solution, is to conditionally dispatch saveBookmarkSuccess action
that mutates the bookmarks state of the current page (remember we override
the bookmark state on different routes).

The way we did this was to dispatch a noOpAction that is dispatched
instead of saveBookmarkSuccess action so the bookmarks state will no
be changed.

Fixes issue: #10